requirements.txt
polygon
streamlit
numpy
scipy
pandas
plotly
python-dotenv
requests
xai-sdk
